# build/swagger.json
{
	"components": {
...
		"schemas": {
			"IEventResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"title": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"startDate": {
						"type": "string",
						"format": "date-time"
					},
					"endDate": {
						"type": "string",
						"format": "date-time"
					},
					"formattedStartDate": {
						"type": "string"
					},
					"formattedEndDate": {
						"type": "string"
					},
					"dayOfWeek": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"title",
					"description",
					"startDate",
					"endDate",
					"formattedStartDate",
					"formattedEndDate",
					"dayOfWeek"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IFreeSlot": {
				"properties": {
					"start": {
						"type": "string"
					},
					"end": {
						"type": "string"
					},
					"day": {
						"type": "string"
					}
				},
				"required": [
					"start",
					"end",
					"day"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICalendar": {
				"properties": {
					"id": {
						"type": "string"
					},
					"name": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEventCreationResponse": {
				"properties": {
					"id": {
						"type": "string"
					}
				},
				"required": [
					"id"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IEventRequestBody": {
				"properties": {
					"summary": {
						"type": "string"
					},
					"location": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"start": {
						"properties": {
							"timeZone": {
								"type": "string"
							},
							"dateTime": {
								"type": "string"
							}
						},
						"required": [
							"timeZone",
							"dateTime"
						],
						"type": "object"
					},
					"end": {
						"properties": {
							"timeZone": {
								"type": "string"
							},
							"dateTime": {
								"type": "string"
							}
						},
						"required": [
							"timeZone",
							"dateTime"
						],
						"type": "object"
					}
				},
				"required": [
					"summary",
					"location",
					"description",
					"start",
					"end"
				],
				"type": "object",
				"additionalProperties": false
			},
			"HandleUserInputRequest": {
				"properties": {
					"userInput": {
						"type": "string"
					},
					"assistantId": {
						"type": "string"
					},
					"threadId": {
						"type": "string"
					}
				},
				"required": [
					"userInput",
					"assistantId",
					"threadId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"GenerateAudioRequest": {
				"properties": {
					"text": {
						"type": "string"
					}
				},
				"required": [
					"text"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "my-task-manager",
		"version": "1.0.0",
		"description": "API service for managing tasks",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
		"/calendar/events": {
			"get": {
				"operationId": "GetEvents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IEventResponse"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "Invalid start or end date"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "start",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "end",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"post": {
				"operationId": "CreateEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IEventCreationResponse"
								}
							}
						}
					},
					"500": {
						"description": "Server Error"
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEventRequestBody"
							}
						}
					}
				}
			}
		},
		"/calendar/free-slots": {
			"get": {
				"operationId": "GetFreeSlots",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IFreeSlot"
									},
									"type": "array"
								}
							}
						}
					},
					"400": {
						"description": "Invalid start or end date"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "start",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "end",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "duration",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/calendar/calendars": {
			"get": {
				"operationId": "GetCalendars",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ICalendar"
									},
									"type": "array"
								}
							}
						}
					},
					"500": {
						"description": "Server Error"
					}
				},
				"security": [],
				"parameters": []
			}
		},
		"/calendar/events/{id}": {
			"put": {
				"operationId": "UpdateEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Event not found"
					},
					"500": {
						"description": "Server Error"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IEventRequestBody"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteEvent",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"404": {
						"description": "Event not found"
					},
					"500": {
						"description": "Server Error"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/calendar/oauth2callback": {
			"get": {
				"operationId": "HandleOAuth2Callback",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"message": {
											"type": "string"
										}
									},
									"required": [
										"message"
									],
									"type": "object"
								}
							}
						}
					},
					"400": {
						"description": "Bad Request"
					},
					"500": {
						"description": "Error retrieving access token"
					}
				},
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "code",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/assistant/user-input": {
			"post": {
				"operationId": "HandleUserInput",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/HandleUserInputRequest"
							}
						}
					}
				}
			}
		},
		"/audio/generate": {
			"post": {
				"operationId": "GenerateAudio",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/GenerateAudioRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "https://sb-api.ngrok.app/"
		}
	]
}